LayoutTests/http/tests/websocket/tests/hybi/binary-type-expected.txt

CONSOLE MESSAGE: line 616: 'Blob' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: line 616: 'ArrayBuffer' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: line 616: '' is not a valid value for binaryType; binaryType remains unchanged.
Test WebSocket.binaryType attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS ws.binaryType is "arraybuffer"
PASS ws.binaryType is "blob"
PASS ws.binaryType = 'Blob' threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS ws.binaryType is "blob"
PASS ws.binaryType = 'ArrayBuffer' threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS ws.binaryType is "blob"
PASS ws.binaryType = '' threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS ws.binaryType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
